+++ app/Http/Controllers/FederationController.php
@@ -131,7 +131,7 @@ class FederationController extends Controller
 			'orderedItems' => []
 		];
 
-		return response(json_encode($res, JSON_UNESCAPED_SLASHES))->header('Content-Type', 'application/ld+json; profile="http://www.w3.org/ns/activitystreams"');
+		return response(json_encode($res, JSON_UNESCAPED_SLASHES))->header('Content-Type', 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"');
 	}
 
 	public function userInbox(Request $request, $username)
@@ -166,17 +166,19 @@ class FederationController extends Controller
 				if(Cache::get($lockKey) !== null) {
 					return;
 				} else {
-					Cache::put($lockKey, 1, 3600);
+					Cache::put($lockKey, 1, 43200);
 					usleep(5000);
 				}
 			}
 			dispatch(new DeleteWorker($headers, $payload))->onQueue('delete');
+		} else if( isset($obj['type']) && in_array($obj['type'], ['Follow', 'Accept'])) {
+			dispatch(new InboxValidator($username, $headers, $payload))->onQueue('follow');
 		} else {
 			$lockKey = 'pf:ap:user-inbox:activity:' . hash('sha256', $obj['id']);
 			if(Cache::get($lockKey) !== null) {
 				return;
 			}
-			Cache::put($lockKey, 1, 3600);
+			Cache::put($lockKey, 1, 43200);
 			usleep(5000);
 			dispatch(new InboxValidator($username, $headers, $payload))->onQueue('high');
 		}
@@ -219,10 +221,12 @@ class FederationController extends Controller
 					return;
 				}
 			}
-			Cache::put($lockKey, 1, 3600);
+			Cache::put($lockKey, 1, 43200);
 			dispatch(new DeleteWorker($headers, $payload))->onQueue('delete');
+		} else if( isset($obj['type']) && in_array($obj['type'], ['Follow', 'Accept'])) {
+			dispatch(new InboxWorker($headers, $payload))->onQueue('follow');
 		} else {
-			dispatch(new InboxWorker($headers, $payload))->onQueue('high');
+			dispatch(new InboxWorker($headers, $payload))->onQueue('shared');
 		}
 		return;
 	}