+++ app/Jobs/InboxPipeline/ActivityHandler.php
@@ -2,32 +2,35 @@
 
 namespace App\Jobs\InboxPipeline;
 
-use App\Util\ActivityPub\Inbox;
 use Illuminate\Bus\Queueable;
+use Illuminate\Contracts\Queue\ShouldBeUnique;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
+use App\Util\ActivityPub\Inbox;
 
-class SharedInboxWorker implements ShouldQueue
+class ActivityHandler implements ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
-    protected $request;
-    protected $profile;
+    protected $username;
+    protected $headers;
     protected $payload;
 
     public $timeout = 300;
     public $tries = 1;
+    public $maxExceptions = 1;
 
     /**
      * Create a new job instance.
      *
      * @return void
      */
-    public function __construct($request, $payload)
+    public function __construct($headers, $username, $payload)
     {
-        $this->request = $request;
+        $this->username = $username;
+        $this->headers = $headers;
         $this->payload = $payload;
     }
 
@@ -38,6 +41,7 @@ class SharedInboxWorker implements ShouldQueue
      */
     public function handle()
     {
-        (new Inbox($this->request, null, $this->payload))->handleSharedInbox();
+        (new Inbox($this->headers, $this->username, $this->payload))->handle();
+        return;
     }
 }