+++ config/logging.php
@@ -89,6 +89,13 @@ return [
         'emergency' => [
             'path' => storage_path('logs/laravel.log'),
         ],
+
+        'media' => [
+            'driver' => 'single',
+            'bubble' => false,
+            'path'   => storage_path('logs/media.log'),
+            'level'  => 'debug'
+        ],
     ],
 
 ];