+++ database/migrations/2022_12_18_012352_add_status_id_index_to_media_table.php
@@ -1,0 +1,36 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('media', function (Blueprint $table) {
+            $table->index('status_id');
+            $table->index('replicated_at');
+            $table->index('version');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('media', function (Blueprint $table) {
+            $table->dropIndex('media_status_id_index');
+            $table->dropIndex('media_replicated_at_index');
+            $table->dropIndex('media_version_index');
+        });
+    }
+};